// src/lib/dsabp.d.ts
export declare abstract class BPCmd {
	/**
	 * @returns A new instance with the same properties. (Deep clone)
	 */
	abstract clone(): any;
	/** @private */
	abstract toArray(): any[];
}
/**
 * Used to specify repeated placements along the X+ axis in a single build command.
 * `101` = place 2 objects with a space between. Cannot exceed 64 bits.
 */
export declare class BuildBits {
	/** @private */
	int: bigint;
	/**
	 * @param input Use to initialize the bits by entering a binary string like `"11001"`.
	 *  Left to right. Trailing zeros are ignored.
	 * @throws SyntaxError if input string cannot be converted to a BigInt.
	 */
	constructor(input?: string | number | bigint);
	/**
	 * Sets a bit at an index (makes it `1`).
	 * @throws RangeError if the index is out of [0,63].
	 */
	set(index: number): this;
	/**
	 * Clears a bit at an index (makes it `0`).
	 * @throws RangeError if the index is out of [0,63].
	 */
	clear(index: number): this;
	/**
	 * Toggles a bit at an index.
	 * @param force If specified, it works as a shortcut for {@link set} (true) and {@link clear} (false).
	 * @throws RangeError if the index is out of [0,63].
	 */
	toggle(index: number, force?: boolean): this;
	/** Checks whether a bit at an index is set. */
	isSet(index: number): boolean;
	/** Checks whether no bit is set. */
	isZero(): boolean;
	/**
	 * Checks whether there is only one bit and it is set to `1`, which is likely
	 * just so that a shape can be specified in the encoded data.
	 */
	isOne(): boolean;
	/**
	 * Removes leading zero bits. (Note that trailing zeros don't work anyway)
	 * @example
	 * const bits = new BuildBits("10100") // 101
	 * bits.del(0) // 001
	 * bits.trimLeadZeros() // 1
	 */
	trimLeadZeros(): this;
	/** @returns Amount of all the bits. */
	get size(): number;
	[Symbol.iterator](): Iterator<boolean>;
	/** @returns Booleans representing the set and unset bits. */
	toArray(): boolean[];
	/** @returns A string of 1's and 0's representing the set and unset bits. */
	toString(): string;
	/** @hidden */
	get [Symbol.toStringTag](): string;
	/** Checks whether this bits equals to another. */
	equals(target: BuildBits): boolean;
	/** @returns A new instance with the same bits. */
	clone(): BuildBits;
}
// ...
export declare class Shape extends Enum<number> {
	readonly vertices: {
		x: number;
		y: number;
	}[];
	constructor(v: number, vertices: {
		x: number;
		y: number;
	}[]);
	/** Checks whether the shape has a flat and full top surface. */
	get isBuildSurface(): boolean;
	/**⣿⣿⣿⣿\
	   ⣿⣿⣿⣿*/
	static BLOCK: Shape;
	/**⣷⣄     \
	   ⣿⣿⣷⣄*/
	static RAMP_UR: Shape;
	/**⣿⣿⡿⠋\
	   ⡿⠋    &#8198;*/
	static RAMP_DR: Shape;
	/**⠙⢿⣿⣿\
	 *       ⠙⢿*/
	static RAMP_DL: Shape;
	/**&#8198;    ⢀⣴\
	   ⢀⣴⣿⣿*/
	static RAMP_UL: Shape;
	/**&#8198;        \
	   ⣿⣿⣿⣿*/
	static SLAB_U: Shape;
	/**⣿⣿     \
	   ⣿⣿    &#8198;*/
	static SLAB_R: Shape;
	/**⣿⣿⣿⣿\
	 *          &#8198;*/
	static SLAB_D: Shape;
	/**&#8198;    ⣿⣿\
	 *       ⣿⣿*/
	static SLAB_L: Shape;
	/**&#8198;        \
	   ⣷⣦⣄⡀*/
	static HALF_RAMP_1_U: Shape;
	/**⣿⠏     \
	   ⠏      &#8198;*/
	static HALF_RAMP_1_R: Shape;
	/**⠈⠙⠻⢿\
	 *          &#8198;*/
	static HALF_RAMP_1_D: Shape;
	/**&#8198;      ⣰\
	 *       ⣰⣿*/
	static HALF_RAMP_1_L: Shape;
	/**⣷⣦⣄⡀\
	   ⣿⣿⣿⣿*/
	static HALF_RAMP_2_U: Shape;
	/**⣿⣿⣿⠏\
	   ⣿⣿⠏  &#8198;*/
	static HALF_RAMP_2_R: Shape;
	/**⣿⣿⣿⣿\
	   ⠈⠙⠻⢿*/
	static HALF_RAMP_2_D: Shape;
	/**&#8198;  ⣰⣿⣿\
	   ⣰⣿⣿⣿*/
	static HALF_RAMP_2_L: Shape;
	/**&#8198;        \
	   ⢀⣠⣴⣾*/
	static HALF_RAMP_1_UI: Shape;
	/**⣆       \
	   ⣿⣆    &#8198;*/
	static HALF_RAMP_1_RI: Shape;
	/**⡿⠟⠋⠁\
	 *          &#8198;*/
	static HALF_RAMP_1_DI: Shape;
	/**&#8198;    ⠹⣿\
	 *         ⠹*/
	static HALF_RAMP_1_LI: Shape;
	/**⢀⣠⣴⣾\
	   ⣿⣿⣿⣿*/
	static HALF_RAMP_2_UI: Shape;
	/**⣿⣿⣆   \
	   ⣿⣿⣿⣆*/
	static HALF_RAMP_2_RI: Shape;
	/**⣿⣿⣿⣿\
	   ⡿⠟⠋⠁*/
	static HALF_RAMP_2_DI: Shape;
	/**⠹⣿⣿⣿\
	 *     ⠹⣿⣿*/
	static HALF_RAMP_2_LI: Shape;
	/**⣷⣦⣄⡀\
	 *          &#8198;*/
	static HALF_RAMP_3_U: Shape;
	/**&#8198;    ⣿⠏\
	 *       ⠏  &#8198;*/
	static HALF_RAMP_3_R: Shape;
	/**&#8198;        \
	   ⠈⠙⠻⢿*/
	static HALF_RAMP_3_D: Shape;
	/**&#8198;  ⣰     \
	   ⣰⣿    &#8198;*/
	static HALF_RAMP_3_L: Shape;
	/**⢀⣠⣴⣾\
	 *          &#8198;*/
	static HALF_RAMP_3_UI: Shape;
	/**&#8198;    ⣆   \
	 *       ⣿⣆*/
	static HALF_RAMP_3_RI: Shape;
	/**&#8198;        \
	   ⡿⠟⠋⠁*/
	static HALF_RAMP_3_DI: Shape;
	/**⠹⣿     \
	 *     ⠹    &#8198;*/
	static HALF_RAMP_3_LI: Shape;
	/**&#8198;        \
	   ⣿⣿    &#8198;*/
	static QUARTER_UR: Shape;
	/**⣿⣿     \
	 *          &#8198;*/
	static QUARTER_DR: Shape;
	/**&#8198;    ⣿⣿\
	 *          &#8198;*/
	static QUARTER_DL: Shape;
	/**&#8198;        \
	 *       ⣿⣿*/
	static QUARTER_UL: Shape;
	/**&#8198;        \
	   ⣷⣄    &#8198;*/
	static QUARTER_RAMP_UR: Shape;
	/**⡿⠋     \
	 *          &#8198;*/
	static QUARTER_RAMP_DR: Shape;
	/**&#8198;    ⠙⢿\
	 *          &#8198;*/
	static QUARTER_RAMP_DL: Shape;
	/**&#8198;        \
	 *       ⣠⣾*/
	static QUARTER_RAMP_UL: Shape;
	/**⣿⣿⣷⣄\
	   ⣿⣿⣿⣿*/
	static BEVEL_UR: Shape;
	/**⣿⣿⣿⣿\
	   ⣿⣿⡿⠋*/
	static BEVEL_DR: Shape;
	/**⣿⣿⣿⣿\
	   ⠙⢿⣿⣿*/
	static BEVEL_DL: Shape;
	/**⢀⣴⣿⣿\
	   ⣿⣿⣿⣿*/
	static BEVEL_UL: Shape;
	/** Set of shapes with a flat and full top surface. */
	static buildSurfaceShapes: Set<Shape>;
}
// ...
export interface DecoderOptions {
	/**
	 * A config command is an array in `[CmdType, null | configMessage]` format.
	 * `null` means that the config command is empty, and the game won't change config of the target object.
	 * `configMessage` contains the config values, and it will be referred to as "config data" in the docs.
	 *
	 * Setting this option to `true` tells the decoder to ignore the config data while decoding a config command, but take the raw data.
	 * The `null` is still read. Whether the config data is not decoded can be checked using {@link ConfigCmd#isRaw | ConfigCmd#isRaw}.
	 *
	 * A {@link Blueprint} containing non-decoded {@link ConfigCmd}s can still be encoded.
	 * The raw data can be decoded individually using the {@link decodeConfigCmd} function, when needed.
	 *
	 * Consider enabling this if you don't need to read all the config commands,
	 * to avoid unnecessary use of resources and decode a little faster.
	 */
	ignoreConfigCmdData?: boolean;
}
/**
 * All the properties can be set, and are required for a valid blueprint.
 */
export interface BlueprintOptions {
	/**
	 * Blueprint format version. 0 by default. At the time of writing, the game only accepts `0` and `-1`
	 * and considers them the same. Any further updates may require changes in the library.
	 * @defaultValue `0`
	 */
	version?: number;
	/**
	 * Width of the blueprint area in square. [1, 100] integer.
	 * @defaultValue `1`
	 */
	width?: number;
	/**
	 * Height of the blueprint area in square. [1, 100] integer.
	 * @defaultValue `1`
	 */
	height?: number;
	/**
	 * All commands of the blueprint.
	 * @example
	 * for (const cmd of bp.commands) {
	 * 	if (cmd instanceof BuildCmd) {
	 * 		console.log(cmd.item.name)
	 * 	}
	 * }
	 * @defaultValue `[]`
	 */
	commands?: Array<BPCmd>;
}
/**
 * All the properties can be set. Set to `undefined` to remove a property. {@link x}, {@link y}, {@link item} are required for a valid command.
 */
export interface BuildCmdOptions {
	/**
	 * X-coord, horizontal offset from the **middle** of the left bottom square of the blueprint area. Can go down to `-0.5`.
	 */
	x?: number;
	/**
	 * Y-coord, vertical offset from the **middle** of left bottom square of the blueprint area. Can go down to `-0.5`.
	 */
	y?: number;
	/** The item to build. See the {@link Item} enum. */
	item?: Item;
	/** A {@link BuildBits} instance. */
	bits?: BuildBits;
	/** Shape of the block to be placed. */
	shape?: Shape;
}
// ...
export interface ConfigCmdOptions {
	/** What the filter for hatches and loaders should do. */
	filterMode?: FilterMode;
	/**  */
	filterItems?: [
		Item,
		Item,
		Item
	];
	/** Used for expando boxes. [0, 360] float. */
	angle?: number;
	/** Used for shield generators. */
	fixedAngle?: FixedAngle;
	/**  */
	pusher?: PusherConfig;
	/**  */
	loader?: LoaderConfig;
}
// ...
export declare class Blueprint implements BlueprintOptions {
	version: BlueprintOptions["version"];
	width: BlueprintOptions["width"];
	height: BlueprintOptions["height"];
	commands: BlueprintOptions["commands"];
	/**
	 * @param input Defaults to a 1x1 blueprint with no commands and version 0.
	 */
	constructor(input?: BlueprintOptions);
	/** Changes multiple properties of the blueprint. */
	set(input: BlueprintOptions): this;
	/** @private */
	fillFromArray(arr: any[], shallow?: boolean): this;
	/** @private */
	toArray(shallow?: boolean): any[];
	/**
	 * @returns A new instance with the same properties. (Deep clone)
	 */
	clone(): Blueprint;
}
export declare class ConfigCmd extends BPCmd implements ConfigCmdOptions {
	/**
	 * An object for default config values, intented to match the game's defaults.
	 * Used for properties that are `null` during encoding.
	 *
	 * You can modify this as you wish, so that you don't have to depend on the defaults of the library.
	 * See the object definition in [ConfigCmd.ts](https://github.com/Blueyescat/dsabp-js/blob/main/src/ConfigCmd.ts).
	 */
	static get defaults(): Required<ConfigCmdOptions>;
	static set defaults(input: Required<ConfigCmdOptions>);
	/** @private */ rawData: Uint8Array;
	filterMode: ConfigCmdOptions["filterMode"];
	filterItems: ConfigCmdOptions["filterItems"];
	angle: ConfigCmdOptions["angle"];
	fixedAngle: ConfigCmdOptions["fixedAngle"];
	pusher: ConfigCmdOptions["pusher"];
	loader: ConfigCmdOptions["loader"];
	/**
	 * @param input If omitted, the config will be empty.
	 * The game won't modify the existing config of an already placed object if the config command is empty.
	 */
	constructor(input?: ConfigCmdOptions);
	/** Changes multiple properties of the command. */
	set(input: ConfigCmdOptions): this;
	/** [cmdType, cfgMsg: null | any[] | Uint8Array] @private */
	fillFromArray(arr: any[]): this;
	/** @private */
	fillDataFromArray(data: any[]): this;
	/** @private */
	toArray(): any[];
	/**
	 * Checks whether the data of this command is not decoded. Comes from decoding a blueprint with
	 * {@link DecoderOptions.ignoreConfigCmdData} = `true`.
	 */
	get isRaw(): boolean;
	/**
	 * Checks whether this command has the same configuration as the target command.
	 */
	equals(target: ConfigCmd): boolean;
	clone(): ConfigCmd;
}
/**
 * Internally created by the {@link decode}, {@link decodeSync}, {@link decodeConfigCmd} and {@link decodeConfigCmdSync} functions.
 *
 * There doesn't seem to be any benefit to using the same instance,
 * but it is possible to do so only with the sync methods.
 */
export declare class Decoder {
	#private;
	/** @private Only exists in browser build. */
	static fflate_inflateSync: (data: Uint8Array, opts?: InflateOptions) => Uint8Array;
	options: DecoderOptions;
	constructor();
	/** {@inheritDoc decodeSync} */
	decodeSync(input: string, options?: DecoderOptions): Blueprint;
	/** {@inheritDoc decodeConfigCmdSync} */
	decodeConfigCmdSync(cmd: ConfigCmd): ConfigCmd;
	/** @private */
	decodeConfigCmdData(rawCmd: Uint8Array): any;
}
// ...
export declare class BuildCmd extends BPCmd implements BuildCmdOptions {
	x: BuildCmdOptions["x"];
	y: BuildCmdOptions["y"];
	item: BuildCmdOptions["item"];
	bits: BuildCmdOptions["bits"];
	shape: BuildCmdOptions["shape"];
	constructor(input?: BuildCmdOptions);
	/** Changes multiple properties of the command. */
	set(input: BuildCmdOptions): this;
	/** @private */
	fillFromArray(arr: any[]): this;
	/** @private */
	toArray(): any[];
	clone(): BuildCmd;
}